+++ Robust.Server/GameStates/PVSSystem.cs
@@ -15,10 +15,8 @@ using Robust.Shared.Timing;
 using Robust.Shared.Utility;
 using System;
 using System.Collections.Generic;
-using System.ComponentModel;
 using System.Linq;
 using System.Runtime.CompilerServices;
-using System.Security.Cryptography;
 using System.Text;
 
 namespace Robust.Server.GameStates;
@@ -323,23 +321,19 @@ internal sealed partial class PVSSystem : EntitySystem
             return;
         DebugTools.Assert(!_mapManager.IsGrid(ev.Sender));
 
-        if (ev.Component.MapUid == ev.Sender)
-            return;
+        if (!ev.Component.ParentUid.IsValid())
+        {
+            // This entity is either a map, terminating, or a rare null-space entity.
+            if (Terminating(ev.Sender))
+                return;
+
+            if (ev.Component.MapUid == ev.Sender)
+                return;
+        }
+
         DebugTools.Assert(!_mapManager.IsMap(ev.Sender));
 
         var xformQuery = GetEntityQuery<TransformComponent>();
-
-        if (ev.ParentChanged)
-        {
-            // If parent changes then the RobustTree for that chunk will no longer be valid and we need to force it as dirty.
-            var oldCoordinates = _transform.GetMoverCoordinates(ev.OldPosition, xformQuery);
-            var indices = PVSCollection<EntityUid>.GetChunkIndices(oldCoordinates.Position);
-            if (ev.Component.GridUid != null)
-                _entityPvsCollection.MarkDirty(new GridChunkLocation(ev.Component.GridUid.Value, indices));
-            else
-                _entityPvsCollection.MarkDirty(new MapChunkLocation(ev.Component.MapID, indices));
-        }
-
         var coordinates = _transform.GetMoverCoordinates(ev.Component, xformQuery);
         UpdateEntityRecursive(ev.Sender, ev.Component, coordinates, xformQuery, false);
     }
@@ -733,13 +727,7 @@ internal sealed partial class PVSSystem : EntitySystem
 
         var deletions = _entityPvsCollection.GetDeletedIndices(fromTick);
 
-        // Iterate over visible chunks and add entities.
-        // TODO reorder chunks and allow early termination?
-        // Previously this iteration terminated early whenever the PVS budget was full. However, if this iteration
-        // iterates over new chunks first, then this can lead to a situation where previously seen entities in currently
-        // still visible chunks are never added to visibleEnts, and therefore get treated as if they had left PVS.
-        // Currently this is fixed by just not exiting the iteration early, but it is probably better to just ensure
-        // that new chunks are always the last to be iterated over.
+        // TODO reorder chunks to prioritize those that are closest to the viewer? Helps make pop0in less visible.
         foreach (var i in chunkIndices)
         {
             var cache = chunkCache[i];
@@ -787,6 +775,10 @@ internal sealed partial class PVSSystem : EntitySystem
 
         foreach (var (uid, visiblity) in visibleEnts)
         {
+            // if an entity is visible, its parents should always be visible.
+            DebugTools.Assert((tQuery.GetComponent(uid).ParentUid is not { Valid: true } parent) || visibleEnts.ContainsKey(parent),
+                $"Attempted to send an entity without sending it's parents. Entity: {ToPrettyString(uid)}.");
+
             if (sessionData.RequestedFull)
             {
                 entityStates.Add(GetFullEntityState(session, uid, mQuery.GetComponent(uid)));
@@ -885,8 +877,10 @@ internal sealed partial class PVSSystem : EntitySystem
             var (entered, shouldAdd) = ProcessEntry(in nodeIndex, lastAcked, lastSent, lastSeen,
                 ref newEntityCount, ref enteredEntityCount, newEntityBudget, enteredEntityBudget);
 
-            if (shouldAdd)
-                AddToSendSet(in nodeIndex, metaDataCache[nodeIndex], toSend, fromTick, entered);
+            if (!shouldAdd)
+                return;
+
+            AddToSendSet(in nodeIndex, metaDataCache[nodeIndex], toSend, fromTick, entered);
         }
 
         var node = tree[nodeIndex];