+++ Robust.Server/Maps/MapLoader.cs
@@ -827,6 +827,198 @@ namespace Robust.Server.Maps
                 GridIDMap.Add(grid.GridEntityId, GridIDMap.Count);
             }
 
+            public YamlNode Serialize()
+            {
+                WriteMetaSection();
+                WriteTileMapSection();
+                PopulateEntityList();
+                WriteGridSection();
+                WriteEntitySection();
+
+                return RootNode.ToYaml();
+            }
+
+            private void WriteMetaSection()
+            {
+                var meta = new MappingDataNode();
+                RootNode.Add("meta", meta);
+                meta.Add("format", MapFormatVersion.ToString(CultureInfo.InvariantCulture));
+                // TODO: Make these values configurable.
+                meta.Add("name", "DemoStation");
+                meta.Add("author", "Space-Wizards");
+
+                //TODO: MapId is null when saveBP is used, another reason this jumbled mess needs to be rewritten
+                var isPostInit = MapId is not null && _mapManager.IsMapInitialized(MapId.Value);
+
+                //TODO: This is a workaround to make SaveBP function
+                foreach (var grid in Grids)
+                {
+                    var mapId = _serverEntityManager.GetComponent<TransformComponent>(grid.GridEntityId).MapID;
+
+                    if (_mapManager.IsMapInitialized(mapId))
+                    {
+                        isPostInit = true;
+                        break;
+                    }
+                }
+
+                meta.Add("postmapinit", isPostInit ? "true" : "false");
+            }
+
+            private void WriteTileMapSection()
+            {
+                var tileMap = new MappingDataNode();
+                RootNode.Add("tilemap", tileMap);
+                foreach (var tileDefinition in _tileDefinitionManager)
+                {
+                    tileMap.Add(tileDefinition.TileId.ToString(CultureInfo.InvariantCulture), tileDefinition.ID);
+                }
+            }
+
+            private void WriteGridSection()
+            {
+                var grids = new SequenceDataNode();
+                RootNode.Add("grids", grids);
+
+                int index = 0;
+                foreach (var grid in Grids)
+                {
+                    _serverEntityManager.GetComponent<MapGridComponent>(grid.GridEntityId).GridIndex = index;
+                    var entry = _serializationManager.WriteValue(grid, context: this);
+                    grids.Add(entry);
+                    index++;
+                }
+            }
+
+            private void PopulateEntityList()
+            {
+                var withoutUid = new HashSet<EntityUid>();
+                var saveCompQuery = _serverEntityManager.GetEntityQuery<MapSaveIdComponent>();
+                var transformCompQuery = _serverEntityManager.GetEntityQuery<TransformComponent>();
+                var metaCompQuery = _serverEntityManager.GetEntityQuery<MetaDataComponent>();
+                foreach (var entity in _serverEntityManager.GetEntities())
+                {
+                    var currentTransform = transformCompQuery.GetComponent(entity);
+                    if (MapId != null && currentTransform.MapID != MapId) continue;
+                    if (MapId == null && (!(currentTransform.GridUid is EntityUid gridId) || !GridIDMap.ContainsKey(gridId))) continue;
+
+                    var currentEntity = entity;
+
+                    // Don't serialize things parented to un savable things.
+                    // For example clothes inside a person.
+                    while (currentEntity.IsValid())
+                    {
+                        if (metaCompQuery.GetComponent(currentEntity).EntityPrototype?.MapSavable == false) break;
+                        currentEntity = transformCompQuery.GetComponent(currentEntity).ParentUid;
+                    }
+
+                    if (currentEntity.IsValid()) continue;
+
+                    Entities.Add(entity);
+
+                    if (!saveCompQuery.TryGetComponent(entity, out var mapSaveComp) ||
+                        !UidEntityMap.TryAdd(mapSaveComp.Uid, entity))
+                    {
+                        // If the id was already saved before, or has no save component we need to find a new id for this entity
+                        withoutUid.Add(entity);
+                    }
+                }
+
+                var uidCounter = 0;
+                foreach (var entity in withoutUid)
+                {
+                    while (UidEntityMap.ContainsKey(uidCounter))
+                    {
+                        // Find next available UID.
+                        uidCounter += 1;
+                    }
+
+                    UidEntityMap.Add(uidCounter, entity);
+                    uidCounter += 1;
+                }
+
+                // Build a reverse lookup
+                EntityUidMap.EnsureCapacity(UidEntityMap.Count);
+                foreach(var (saveId, mapId) in UidEntityMap)
+                {
+                    EntityUidMap.Add(mapId, saveId);
+                }
+            }
+
+            private void WriteEntitySection()
+            {
+                var serializationManager = IoCManager.Resolve<ISerializationManager>();
+                var compFactory = IoCManager.Resolve<IComponentFactory>();
+                var metaQuery = _serverEntityManager.GetEntityQuery<MetaDataComponent>();
+                var entities = new SequenceDataNode();
+                RootNode.Add("entities", entities);
+
+                var prototypeCompCache = new Dictionary<string, Dictionary<string, MappingDataNode>>();
+                foreach (var (saveId, entityUid) in UidEntityMap.OrderBy(e=>e.Key))
+                {
+                    CurrentWritingEntity = entityUid;
+                    var mapping = new MappingDataNode
+                    {
+                        {"uid", saveId.ToString(CultureInfo.InvariantCulture)}
+                    };
+
+                    var md = metaQuery.GetComponent(entityUid);
+
+                    Dictionary<string, MappingDataNode>? cache = null;
+                    if (md.EntityPrototype is {} prototype)
+                    {
+                        mapping.Add("type", prototype.ID);
+                        if (!prototypeCompCache.TryGetValue(prototype.ID, out cache))
+                        {
+                            prototypeCompCache[prototype.ID] = cache =  new Dictionary<string, MappingDataNode>();
+                            foreach (var (compType, comp) in prototype.Components)
+                            {
+                                cache.Add(compType, serializationManager.WriteValueAs<MappingDataNode>(comp.Component.GetType(), comp.Component));
+                            }
+                        }
+                    }
+
+                    var components = new SequenceDataNode();
+
+                    // See engine#636 for why the Distinct() call.
+                    foreach (var component in _serverEntityManager.GetComponents(entityUid))
+                    {
+                        if (component is MapSaveIdComponent)
+                            continue;
+
+                        var compType = component.GetType();
+                        var compName = compFactory.GetComponentName(compType);
+                        CurrentWritingComponent = compName;
+                        var compMapping = serializationManager.WriteValueAs<MappingDataNode>(compType, component, context: this);
+
+                        if (cache != null && cache.TryGetValue(compName, out var protMapping))
+                        {
+                            // This will NOT recursively call Except() on the values of the mapping. It will only remove
+                            // key-value pairs if both the keys and values are equal.
+                            compMapping = compMapping.Except(protMapping);
+                            if(compMapping == null) continue;
+                        }
+
+                        // Don't need to write it if nothing was written! Note that if this entity has no associated
+                        // prototype, we ALWAYS want to write the component, because merely the fact that it exists is
+                        // information that needs to be written.
+                        if (compMapping.Children.Count != 0 || md.EntityPrototype == null)
+                        {
+                            compMapping.Add("type", new ValueDataNode(compName));
+                            // Something actually got written!
+                            components.Add(compMapping);
+                        }
+                    }
+
+                    if (components.Count != 0)
+                    {
+                        mapping.Add("components", components);
+                    }
+
+                    entities.Add(mapping);
+                }
+            }
+
             // Create custom object serializers that will correctly allow data to be overriden by the map file.
             MappingDataNode IEntityLoadContext.GetComponentData(string componentName,
                 MappingDataNode? protoData)