+++ Robust.Server/Physics/GridFixtureSystem.cs
@@ -247,18 +247,19 @@ namespace Robust.Server.Physics
                 var mapBody = bodyQuery.GetComponent(mapGrid.GridEntityId);
                 var oldGridComp = gridQuery.GetComponent(mapGrid.GridEntityId);
                 var newGrids = new EntityUid[grids.Count - 1];
+                var mapId = oldGridXform.MapID;
 
                 for (var i = 0; i < grids.Count - 1; i++)
                 {
                     var group = grids[i];
-                    var splitGrid = _mapManager.CreateGrid(mapGrid.ParentMapId);
+                    var splitGrid = _mapManager.CreateGrid(mapId);
+                    var splitXform = xformQuery.GetComponent(splitGrid.GridEntityId);
                     newGrids[i] = splitGrid.GridEntityId;
 
                     // Keep same origin / velocity etc; this makes updating a lot faster and easier.
-                    splitGrid.WorldPosition = gridPos;
-                    splitGrid.WorldRotation = gridRot;
+                    splitXform.WorldPosition = gridPos;
+                    splitXform.WorldRotation = gridRot;
                     var splitBody = bodyQuery.GetComponent(splitGrid.GridEntityId);
-                    var splitXform = xformQuery.GetComponent(splitGrid.GridEntityId);
                     splitBody.LinearVelocity = mapBody.LinearVelocity;
                     splitBody.AngularVelocity = mapBody.AngularVelocity;