+++ Robust.Shared/Containers/Container.cs
@@ -37,36 +37,39 @@ namespace Robust.Shared.Containers
         public override string ContainerType => ClassName;
 
         /// <inheritdoc />
-        protected override void InternalInsert(EntityUid toinsert, EntityUid oldParent, IEntityManager entMan)
+        protected override void InternalInsert(EntityUid toInsert, IEntityManager entMan)
         {
-            // Why TF is this even a list??????
-            DebugTools.Assert(!_containerList.Contains(toinsert));
-            _containerList.Add(toinsert);
-            base.InternalInsert(toinsert, oldParent, entMan);
+            DebugTools.Assert(!_containerList.Contains(toInsert));
+            _containerList.Add(toInsert);
         }
 
         /// <inheritdoc />
-        protected override void InternalRemove(EntityUid toremove, IEntityManager entMan, MetaDataComponent? meta = null)
+        protected override void InternalRemove(EntityUid toRemove, IEntityManager entMan)
         {
-            _containerList.Remove(toremove);
-            base.InternalRemove(toremove, entMan, meta);
+            _containerList.Remove(toRemove);
         }
 
         /// <inheritdoc />
         public override bool Contains(EntityUid contained)
         {
-            return _containerList.Contains(contained);
+            if (!_containerList.Contains(contained))
+                return false;
+
+            var flags = IoCManager.Resolve<IEntityManager>().GetComponent<MetaDataComponent>(contained).Flags;
+            DebugTools.Assert((flags & MetaDataFlags.InContainer) != 0);
+
+            return true;
         }
 
         /// <inheritdoc />
-        public override void Shutdown()
+        protected override void InternalShutdown(IEntityManager entMan, bool isClient)
         {
-            base.Shutdown();
-
-            var entMan = IoCManager.Resolve<IEntityManager>();
-            foreach (var entity in _containerList)
+            foreach (var entity in _containerList.ToArray())
             {
-                entMan.DeleteEntity(entity);
+                if (!isClient)
+                    entMan.DeleteEntity(entity);
+                else if (entMan.EntityExists(entity))
+                    Remove(entity, entMan, reparent: false, addToBroadphase: true, force: true);
             }
         }
     }