+++ Robust.Shared/Physics/Events/CollisionChangeEvent.cs
@@ -3,6 +3,9 @@ using Robust.Shared.Physics.Components;
 
 namespace Robust.Shared.Physics.Events
 {
+    /// <summary>
+    ///     These events are broadcast (not directed) whenever an entity's ability to collide changes.
+    /// </summary>
     [ByRefEvent]
     public readonly struct CollisionChangeEvent
     {