+++ Robust.Shared/Physics/Systems/SharedBroadphaseSystem.cs
@@ -11,6 +11,7 @@ using Robust.Shared.Physics.Components;
 using Robust.Shared.Physics.Dynamics;
 using Robust.Shared.Physics.Dynamics.Contacts;
 using Robust.Shared.Utility;
+using SharpZstd.Interop;
 
 namespace Robust.Shared.Physics.Systems
 {
@@ -75,7 +76,8 @@ namespace Robust.Shared.Physics.Systems
             MapId mapId,
             HashSet<IMapGrid> movedGrids,
             Dictionary<FixtureProxy, Box2> gridMoveBuffer,
-            EntityQuery<BroadphaseComponent> broadQuery)
+            EntityQuery<BroadphaseComponent> broadQuery,
+            EntityQuery<TransformComponent> xformQuery)
         {
             // None moved this tick
             if (movedGrids.Count == 0) return;
@@ -88,8 +90,10 @@ namespace Robust.Shared.Physics.Systems
 
             foreach (var grid in movedGrids)
             {
-                DebugTools.Assert(grid.ParentMapId == mapId);
-                var worldAABB = grid.WorldAABB;
+                var xform = xformQuery.GetComponent(grid.GridEntityId);
+
+                DebugTools.Assert(xform.MapID == mapId);
+                var worldAABB = xform.WorldMatrix.TransformBox(grid.LocalAABB);
                 var enlargedAABB = worldAABB.Enlarged(_broadphaseExpand);
                 var state = (moveBuffer, gridMoveBuffer);
 
@@ -149,7 +153,7 @@ namespace Robust.Shared.Physics.Systems
             var xformQuery = GetEntityQuery<TransformComponent>();
 
             // Find any entities being driven over that might need to be considered
-            FindGridContacts(component, mapId, movedGrids, gridMoveBuffer, broadphaseQuery);
+            FindGridContacts(component, mapId, movedGrids, gridMoveBuffer, broadphaseQuery, xformQuery);
 
             // There is some mariana trench levels of bullshit going on.
             // We essentially need to re-create Box2D's FindNewContacts but in a way that allows us to check every
@@ -204,8 +208,8 @@ namespace Robust.Shared.Physics.Systems
                     if (proxyA.Fixture.Hard && other.Fixture.Hard &&
                         (gridMoveBuffer.ContainsKey(proxyA) || gridMoveBuffer.ContainsKey(other)))
                     {
-                        _physicsSystem.WakeBody(proxyABody);
-                        _physicsSystem.WakeBody(otherBody);
+                        _physicsSystem.WakeBody(proxyABody, force: true);
+                        _physicsSystem.WakeBody(otherBody, force: true);
                     }
 
                     contactManager.AddPair(proxyA, other);
@@ -233,10 +237,10 @@ namespace Robust.Shared.Physics.Systems
 
             foreach (var grid in movedGrids)
             {
-                DebugTools.Assert(grid.ParentMapId == mapId);
+                var xform = xformQuery.GetComponent(grid.GridEntityId);
+                DebugTools.Assert(xform.MapID == mapId);
 
                 var mapGrid = (MapGrid)grid;
-                var xform = xformQuery.GetComponent(grid.GridEntityId);
 
                 var (worldPos, worldRot, worldMatrix, invWorldMatrix) = xform.GetWorldPositionRotationMatrixWithInv(xformQuery);
 
@@ -250,11 +254,15 @@ namespace Robust.Shared.Physics.Systems
 
                 foreach (var colliding in _mapManager.FindGridsIntersecting(mapId, aabb, gridsPool, xformQuery, bodyQuery, true))
                 {
-                    if (grid == colliding) continue;
+                    if (grid == colliding ||
+                        !xformQuery.TryGetComponent(colliding.GridEntityId, out var collidingXform))
+                    {
+                        continue;
+                    }
 
-                    var otherGrid = (MapGrid)colliding;
-                    var otherGridBounds = colliding.WorldAABB;
-                    var otherGridInvMatrix = colliding.InvWorldMatrix;
+                    var otherGrid = (MapGrid) colliding;
+                    var (_, _, otherGridMatrix, otherGridInvMatrix) =  collidingXform.GetWorldPositionRotationMatrixWithInv();
+                    var otherGridBounds = otherGridMatrix.TransformBox(colliding.LocalAABB);
                     var otherTransform = _physicsSystem.GetPhysicsTransform(colliding.GridEntityId, xformQuery: xformQuery);
 
                     // Get Grid2 AABB in grid1 ref
@@ -319,9 +327,7 @@ namespace Robust.Shared.Physics.Systems
 
             // Logger.DebugS("physics", $"Checking proxy for {proxy.Fixture.Body.Owner} on {broadphase.Owner}");
             Box2 aabb;
-            var proxyBroad = _lookup.GetBroadphase(proxy.Fixture.Body.Owner, xform, broadphaseQuery, xformQuery);
-
-            if (proxyBroad == null)
+            if (!_lookup.TryGetCurrentBroadphase(xform, out var proxyBroad))
             {
                 _logger.Error($"Found null broadphase for {ToPrettyString(proxy.Fixture.Body.Owner)}");
                 DebugTools.Assert(false);
@@ -387,22 +393,18 @@ namespace Robust.Shared.Physics.Systems
             }, aabb, true);
         }
 
-        public void RegenerateContacts(PhysicsComponent body)
+        public void RegenerateContacts(PhysicsComponent body, TransformComponent? xform = null, FixturesComponent? fixtures = null)
         {
             _physicsSystem.DestroyContacts(body);
-            var broadQuery = GetEntityQuery<BroadphaseComponent>();
-            var xformQuery = GetEntityQuery<TransformComponent>();
+            if (!Resolve(body.Owner, ref xform, ref fixtures))
+                return;
 
-            var broadphase = _lookup.GetBroadphase(body.Owner, xformQuery.GetComponent(body.Owner), broadQuery, xformQuery);
+            if (!_lookup.TryGetCurrentBroadphase(xform, out var broadphase))
+                return;
 
-            if (broadphase != null)
+            foreach (var fixture in fixtures.Fixtures.Values)
             {
-                var mapId = EntityManager.GetComponent<TransformComponent>(body.Owner).MapID;
-
-                foreach (var fixture in EntityManager.GetComponent<FixturesComponent>(body.Owner).Fixtures.Values)
-                {
-                    TouchProxies(mapId, broadphase, fixture);
-                }
+                TouchProxies(xform.MapID, broadphase, fixture);
             }
         }
 
@@ -426,7 +428,7 @@ namespace Robust.Shared.Physics.Systems
             physicsMap.MoveBuffer[proxy] = aabb;
         }
 
-        public void Refilter(Fixture fixture)
+        public void Refilter(Fixture fixture, TransformComponent? xform = null)
         {
             // TODO: Call this method whenever collisionmask / collisionlayer changes
             // TODO: This should never becalled when body is null.
@@ -441,14 +443,13 @@ namespace Robust.Shared.Physics.Systems
                 contact.Flags |= ContactFlags.Filter;
             }
 
-            var broadQuery = GetEntityQuery<BroadphaseComponent>();
-            var xformQuery = GetEntityQuery<TransformComponent>();
-            var broadphase = _lookup.GetBroadphase(fixture.Body.Owner, xformQuery.GetComponent(fixture.Body.Owner), broadQuery, xformQuery);
+            if (!Resolve(fixture.Body.Owner, ref xform))
+                return;
 
-            // If nullspace or whatever ignore it.
-            if (broadphase == null) return;
+            if (!_lookup.TryGetCurrentBroadphase(xform, out var broadphase))
+                return;
 
-            TouchProxies(Transform(fixture.Body.Owner).MapID, broadphase, fixture);
+            TouchProxies(xform.MapID, broadphase, fixture);
         }
 
         // TODO: The below is slow and should just query the map's broadphase directly. The problem is that