+++ Robust.UnitTesting/Server/GameObjects/Components/Container_Test.cs
@@ -6,6 +6,7 @@ using Robust.Shared.Containers;
 using Robust.Shared.GameObjects;
 using Robust.Shared.IoC;
 using Robust.Shared.Map;
+using Robust.Shared.Utility;
 
 // ReSharper disable AccessToStaticMemberViaDerivedType
 
@@ -305,35 +306,37 @@ namespace Robust.UnitTesting.Server.GameObjects.Components
             public override List<EntityUid> ExpectedEntities => _expectedEntities;
 
             /// <inheritdoc />
-            protected override void InternalInsert(EntityUid toinsert, EntityUid oldParent, IEntityManager entMan)
+            protected override void InternalInsert(EntityUid toInsert, IEntityManager entMan)
             {
-                _containerList.Add(toinsert);
-                base.InternalInsert(toinsert, oldParent, entMan);
+                _containerList.Add(toInsert);
             }
 
             /// <inheritdoc />
-            protected override void InternalRemove(EntityUid toremove, IEntityManager entMan, MetaDataComponent? meta = null)
+            protected override void InternalRemove(EntityUid toRemove, IEntityManager entMan)
             {
-                _containerList.Remove(toremove);
-                base.InternalRemove(toremove, entMan, meta);
+                _containerList.Remove(toRemove);
             }
 
             /// <inheritdoc />
             public override bool Contains(EntityUid contained)
             {
-                return _containerList.Contains(contained);
+                if (!_containerList.Contains(contained))
+                    return false;
+
+                var flags = IoCManager.Resolve<IEntityManager>().GetComponent<MetaDataComponent>(contained).Flags;
+                DebugTools.Assert((flags & MetaDataFlags.InContainer) != 0);
+                return true;
             }
 
             /// <inheritdoc />
-            public override void Shutdown()
+            protected override void InternalShutdown(IEntityManager entMan, bool isClient)
             {
-                base.Shutdown();
-
-                var entMan = IoCManager.Resolve<IEntityManager>();
-
-                foreach (var entity in _containerList)
+                foreach (var entity in _containerList.ToArray())
                 {
-                    entMan.DeleteEntity(entity);
+                    if (!isClient)
+                        entMan.DeleteEntity(entity);
+                    else if (entMan.EntityExists(entity))
+                        Remove(entity, entMan, reparent: false, addToBroadphase: true, force: true);
                 }
             }